test(dev): migrate hmr tests using new `dev` API
