+++ crates/rolldown/src/dev/build_driver.rs
@@ -2,6 +2,7 @@ use std::{mem, path::PathBuf, sync::Arc};
 
 use futures::FutureExt;
 
+use rolldown_common::HmrUpdate;
 use rolldown_error::BuildResult;
 use tokio::sync::Mutex;
 
@@ -111,7 +112,7 @@ impl BuildDriver {
     &self,
     caller: String,
     first_invalidated_by: Option<String>,
-  ) -> BuildResult<()> {
+  ) -> BuildResult<HmrUpdate> {
     let mut build_state = loop {
       let build_state = self.ctx.state.lock().await;
       if let Some(building_future) = build_state.is_busy_then_future().cloned() {
@@ -125,12 +126,9 @@ impl BuildDriver {
     let bundler = self.bundler.lock().await;
     let cache = build_state.cache.take().expect("Should never be none here");
     let mut hmr_manager = bundler.create_hmr_manager(cache);
-    let updates =
+    let update =
       hmr_manager.compute_update_for_calling_invalidate(caller, first_invalidated_by).await?;
     build_state.cache = Some(hmr_manager.input.cache);
-    if let Some(on_hmr_updates) = self.ctx.options.on_hmr_updates.as_ref() {
-      on_hmr_updates(vec![updates]);
-    }
-    Ok(())
+    Ok(update)
   }
 }
