+++ crates/rolldown/src/dev/dev_engine.rs
@@ -2,6 +2,7 @@ use std::{ops::Deref, sync::Arc};
 
 use arcstr::ArcStr;
 use futures::{FutureExt, future::Shared};
+use rolldown_common::HmrUpdate;
 use rolldown_error::BuildResult;
 use rolldown_utils::dashmap::FxDashSet;
 use rolldown_watcher::Watcher;
@@ -117,9 +118,8 @@ impl<W: Watcher + Send + 'static> DevEngine<W> {
     &self,
     caller: String,
     first_invalidated_by: Option<String>,
-  ) -> BuildResult<()> {
-    self.build_driver.invalidate(caller, first_invalidated_by).await?;
-    Ok(())
+  ) -> BuildResult<HmrUpdate> {
+    self.build_driver.invalidate(caller, first_invalidated_by).await
   }
 }
