+++ packages/test-dev-server/src/dev-server.ts
@@ -1,126 +1,123 @@
-import chokidar from 'chokidar';
 import connect from 'connect';
 import nodeFs from 'node:fs';
 import http from 'node:http';
 import nodePath from 'node:path';
 import nodeUrl from 'node:url';
 import * as rolldown from 'rolldown';
+import { dev, DevEngine } from 'rolldown/experimental';
 import serveStatic from 'serve-static';
 import { WebSocket, WebSocketServer } from 'ws';
 import { HmrInvalidateMessage } from './types/client-message.js';
 import { NormalizedDevOptions } from './types/normalized-dev-options.js';
 import { HmrUpdateMessage } from './types/server-message.js';
+import { createDevServerPlugin } from './utils/create-dev-server-plugin.js';
 import { decodeClientMessage } from './utils/decode-client-message.js';
+import { loadDevConfig } from './utils/load-dev-config.js';
+import { normalizeDevOptions } from './utils/normalize-dev-options.js';
 
 let seed = 0;
 
-export class DevServer {
-  private numberOfLiveConnections = 0;
-  private sockets: Set<WebSocket> = new Set();
-  private currentBuildingPromise: Promise<rolldown.RolldownOutput> | null =
-    null;
+class DevServer {
+  connectServer = connect();
+  server = http.createServer(this.connectServer);
+  serverStatus = {
+    allowRequest: false,
+    allowRequestPromiseResolvers: Promise.withResolvers<void>(),
+  };
+  wsServer = new WebSocketServer({ server: this.server });
+  #sockets = new Set<WebSocket>();
+  #devOptions?: NormalizedDevOptions;
+  #devEngine?: DevEngine;
 
-  constructor(
-    private buildOptions: rolldown.BuildOptions,
-    private devOptions: NormalizedDevOptions,
-    private connectServer: connect.Server,
-    private httpServer: http.Server,
-    private wsServer: WebSocketServer,
-    private build: rolldown.RolldownBuild,
-  ) {
-  }
+  constructor() {}
 
   async serve(): Promise<void> {
+    this.#prepareServer();
+
+    const devConfig = await loadDevConfig();
+    const devOptions = normalizeDevOptions(devConfig.dev ?? {});
+    this.#devOptions = devOptions;
+    const buildOptions = devConfig.build ?? {};
+    if (buildOptions.plugins == null || Array.isArray(buildOptions.plugins)) {
+      buildOptions.plugins = [
+        ...(buildOptions.plugins || []),
+        createDevServerPlugin(devOptions),
+      ];
+    } else {
+      throw new Error('Plugins must be an array');
+    }
+
+    const { output: outputOptions, ...inputOptions } = buildOptions;
+    let devEngine = await dev(inputOptions, outputOptions ?? {}, {
+      onHmrUpdates: (updates) => {
+        this.handleHmrUpdates(updates);
+      },
+    });
+    this.#devEngine = devEngine;
+    this.#prepareHttpServerAfterCreateDevEngine(devEngine);
+    await devEngine.run();
+    this.#readyHttpServer();
+  }
+
+  #prepareServer(): void {
+    this.connectServer.use(async (_req, _res, next) => {
+      if (this.serverStatus.allowRequest) {
+        next();
+      } else {
+        await this.serverStatus.allowRequestPromiseResolvers.promise;
+        next();
+      }
+    });
+
     this.wsServer.on('connection', (ws, _req) => {
-      this.numberOfLiveConnections += 1;
-      console.debug(
-        `Detected new Websocket connection. Current live connections: ${this.numberOfLiveConnections}`,
-      );
-      this.sockets.add(ws);
+      this.#sockets.add(ws);
       ws.on('error', console.error);
       ws.on('close', () => {
-        this.numberOfLiveConnections -= 1;
-        console.debug(
-          `Websocket connection closed. Current live connections: ${this.numberOfLiveConnections}`,
-        );
+        // TODO: handle close
       });
-      ws.on('message', (rawData) => {
+      ws.on('message', async (rawData) => {
         const clientMessage = decodeClientMessage(rawData);
         switch (clientMessage.type) {
           case 'hmr:invalidate':
-            this.handleHmrInvalidate(this.build, clientMessage);
+            await this.#handleHmrInvalidate(clientMessage);
             break;
         }
       });
     });
-    const initialBuildPromise = this.triggerBuild();
-    if (this.devOptions.platform === 'browser') {
-      this.connectServer.use(async function(req, _res, next) {
-        if (req.url === '/' || req.url === '/index.html') {
-          console.info('Detected requests. Waiting for initial build...');
-          await initialBuildPromise;
-          next();
-        } else {
-          next();
-        }
-      });
-      this.connectServer.use(
-        serveStatic(nodePath.join(process.cwd(), 'dist'), {
-          index: ['index.html'],
-          extensions: ['html'],
-        }),
-      );
-    }
 
-    this.httpServer.listen(3000, () => {
-      console.log(`Serving ${nodePath.join(process.cwd(), 'dist')}`);
+    this.server.listen(3000, () => {
       console.log('Server listening on http://localhost:3000');
     });
+  }
 
-    const watcher = chokidar.watch(nodePath.join(process.cwd(), 'src'), {
-      usePolling: true,
-      interval: 100,
-    });
-
-    watcher.on('change', async (path) => {
-      console.log(`File ${path} has been changed`);
-      if (this.hasLiveConnections) {
-        const updates = (await this.build.generateHmrPatch([path]))!;
-        this.handleHmrUpdates(updates);
+  #prepareHttpServerAfterCreateDevEngine(devEngine: DevEngine): void {
+    this.connectServer.use(async (req, _res, next) => {
+      if (req.url === '/' || req.url === '/index.html') {
+        await devEngine.ensureLatestBuild();
+        next();
+      } else {
+        next();
       }
-      // Invoke the build process again to ensure if users reload the page, they get the latest changes
-      await this.triggerBuild();
     });
-  }
-
-  get hasLiveConnections(): boolean {
-    return this.numberOfLiveConnections > 0;
-  }
-
-  async triggerBuild(): Promise<void> {
-    if (this.currentBuildingPromise != null) {
-      await this.currentBuildingPromise;
-    }
-    const buildingPromise = this.build.write(this.buildOptions.output).finally(
-      () => {
-        if (this.currentBuildingPromise === buildingPromise) {
-          this.currentBuildingPromise = null;
-        }
-      },
+    this.connectServer.use(
+      serveStatic(nodePath.join(process.cwd(), 'dist'), {
+        index: ['index.html'],
+        extensions: ['html'],
+      }),
     );
-    this.currentBuildingPromise = buildingPromise;
-    await buildingPromise;
   }
 
-  get isInBuilding(): boolean {
-    return this.currentBuildingPromise != null;
+  #readyHttpServer() {
+    this.serverStatus.allowRequest = true;
+    this.serverStatus.allowRequestPromiseResolvers.resolve();
   }
 
-  sendMessage(message: HmrUpdateMessage): void {
-    if (this.hasLiveConnections) {
-      for (const s of this.sockets) {
+  #sendMessage(message: HmrUpdateMessage): void {
+    if (this.#sockets.size > 0) {
+      const encoded = JSON.stringify(message);
+      for (const s of this.#sockets) {
         if (s.readyState === WebSocket.OPEN) {
-          s.send(JSON.stringify(message));
+          s.send(encoded);
         }
       }
     }
@@ -135,7 +132,7 @@ export class DevServer {
           this.sendUpdateToClient(update);
           break;
         case 'FullReload':
-          if (this.devOptions.platform === 'browser') {
+          if (this.#devOptions?.platform === 'browser') {
             // TODO: send reload message to client
           }
           break;
@@ -153,7 +150,7 @@ export class DevServer {
     if (output.type !== 'Patch') {
       return;
     }
-    if (this.hasLiveConnections && output.code) {
+    if (output.code) {
       console.log('Patching...');
       const path = `${seed}.js`;
       seed++;
@@ -173,7 +170,7 @@ export class DevServer {
           path: patchUriForFile,
         }),
       );
-      this.sendMessage({
+      this.#sendMessage({
         type: 'hmr:update',
         url: patchUriForBrowser,
         path: patchUriForFile,
@@ -183,7 +180,6 @@ export class DevServer {
         `Failed to send update to client due to ${
           JSON.stringify(
             {
-              hasLiveConnections: this.hasLiveConnections,
               hasCode: output.code != null,
             },
             null,
@@ -194,14 +190,18 @@ export class DevServer {
     }
   }
 
-  async handleHmrInvalidate(
-    build: rolldown.RolldownBuild,
+  async #handleHmrInvalidate(
     msg: HmrInvalidateMessage,
   ): Promise<void> {
     console.log('Invalidating...');
-    if (this.hasLiveConnections) {
-      const output = await build.hmrInvalidate(msg.moduleId);
+    if (this.#sockets.size > 0) {
+      const output = await this.#devEngine!.invalidate(msg.moduleId);
       this.sendUpdateToClient(output);
     }
   }
+}
+
+export async function serve(): Promise<void> {
+  const devServer = new DevServer();
+  await devServer.serve();
 }
