+++ packages/test-dev-server/src/index.ts
@@ -1,5 +1,5 @@
-import { serveNew } from './dev-server-new.js';
-import { serve } from './serve.js';
+import { serve } from './dev-server.js';
+import { serveLegacy } from './serve.js';
 import { defineDevConfig } from './utils/define-dev-config.js';
 
-export { defineDevConfig, serve, serveNew };
+export { defineDevConfig, serve, serveLegacy };
