+++ packages/test-dev-server/tests/fixtures/invalidation/dev.config.mjs
@@ -5,7 +5,9 @@ export default defineDevConfig({
   build: {
     input: 'src/main.js',
     experimental: {
-      hmr: true,
+      hmr: {
+        new: true,
+      },
     },
     platform: 'node',
     treeshake: false,
